change connection/id/update to connection/id for updating
